+++ nistbeacon/__init__.py
@@ -17,5 +17,9 @@ limitations under the License.
 from .nistbeacon import NistBeacon
 from .nistbeaconvalue import NistBeaconValue
 
-__all__ = ['NistBeacon', 'NistBeaconValue']
-__version__ = '0.9.0'
+__all__ = [
+    'NistBeacon',
+    'NistBeaconValue',
+]
+
+__version__ = '0.9.1'