+++ nistbeacon/nistbeaconvalue.py
@@ -21,11 +21,8 @@ import struct
 from random import Random
 from xml.etree import ElementTree
 
-from Crypto.Hash import SHA512
-from Crypto.PublicKey import RSA
-from Crypto.Signature import PKCS1_v1_5
-
 import nistbeacon.constants as cn
+from nistbeacon.nistbeaconcrypto import NistBeaconCrypto
 
 
 class NistBeaconValue(object):
@@ -34,10 +31,6 @@ class NistBeaconValue(object):
     It has all the normal properties of a NIST beacon API call,
     but stored as a python object
     """
-
-    _understood_namespaces = {
-        'nist-0.1': 'http://beacon.nist.gov/record/0.1/',
-    }
 
     def __init__(
             self,
@@ -125,7 +118,7 @@ class NistBeaconValue(object):
         )
 
         # Signature checking
-        rsa_message = SHA512.new(
+        sha512_hash = NistBeaconCrypto.get_hash(
             self.version.encode() +
             struct.pack(
                 '>1I1Q64s64s1I',
@@ -137,13 +130,8 @@ class NistBeaconValue(object):
             )
         )
 
-        # Get the RSA key, and build a verifier with it
-        rsa_key = RSA.importKey(cn.NIST_RSA_KEY)
-        verifier = PKCS1_v1_5.new(rsa_key)
-
-        # Check the message against the signature
-        sig_check_result = verifier.verify(
-            rsa_message,
+        sig_check_result = NistBeaconCrypto.verify(
+            sha512_hash,
             binascii.a2b_hex(self.signature_value)[::-1]
         )
 
@@ -372,6 +360,10 @@ class NistBeaconValue(object):
 
         invalid_result = None
 
+        understood_namespaces = {
+            'nist-0.1': 'http://beacon.nist.gov/record/0.1/',
+        }
+
         # Our required values are "must haves". This makes it simple
         # to verify we loaded everything out of XML correctly.
         required_values = {
@@ -393,22 +385,16 @@ class NistBeaconValue(object):
 
         # Using the required values, let's load the xml values in
         for key in required_values:
-            discovered_element = tree.find(key)
+            discovered_element = tree.find(
+                "{0}:{1}".format('nist-0.1', key),
+                namespaces=understood_namespaces,
+            )
 
-            if discovered_element is None:
-                # BUG FIX the namespace added by NIST
-                # Issue: https://github.com/urda/nistbeacon/issues/8
+            if not isinstance(discovered_element, ElementTree.Element):
+                continue
 
-                discovered_element = tree.find(
-                    "{0}{1}".format(
-                        '{http://beacon.nist.gov/record/0.1/}',
-                        key,
-                    )
-                )
-
-                if discovered_element is None:
-                    continue
-
+            # Bad pylint message - https://github.com/PyCQA/pylint/issues/476
+            # pylint: disable=no-member
             required_values[key] = discovered_element.text
 
         # Confirm that the required values are set, and not 'None'